requirements.txt
python-dotenv
chromadb
sentence-transformers
openai
yara-python
streamlit
pandas
